Refactorize monitoring module model
